api/requirements.txt
flask
flask_cors
load_dotenv
numpy
pandas
python-multipart